refactor: Junit5 parameterized test 테스트명 수정 및 패키지 분리
