+++ src/test/java/baseball/testlibrary/AssertjTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.testlibrary;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
